♻️#143 Refactor: ExamCommandService 정리 get 메서드는 ExamQueryService로 이동 getExamResult(), getExamIncorrectChoices()
